Remove SystemOperationTest, add docstrings to source function
