+++ tests/source.py
@@ -2,7 +2,6 @@
 
 __author__ = 'Boris Polyanskiy'
 
-from abc import ABCMeta, abstractmethod
 from io import StringIO
 import glob
 from os import listdir, remove
@@ -12,12 +11,11 @@ import subprocess
 import sys
 import unittest
 
-from .generate_test_data import generate_data
-
 GLOBS = ('plot_slice_*', 'plot_axis_*', 'plot_rotational_*', 'plot_zero_*', '*cut_1536', '*step_*40_*', '*step_*100_*')
 
 
 def clean(*args):
+    """Deletes all files and folders in args"""
     for arg in args:
         if isfile(arg):
             remove(arg)
@@ -26,6 +24,7 @@ def clean(*args):
 
 
 def stdout_wrapper(func, *args, **kwargs):
+    """Call func with redirected stdout and return output."""
     with StringIO() as log:
         tmp, sys.stdout = sys.stdout, log
         func(*args, **kwargs)
@@ -34,14 +33,17 @@ def stdout_wrapper(func, *args, **kwargs):
 
 
 def check_glob(src, dst, data):
+    """Check dst folder exists and listdir in dst equals with data."""
     return isdir(dst) and sorted(data) == sorted(join(src, x) for x in listdir(dst))
 
 
 def check_not_glob(src, data):
+    """Check each element in data not in src glob."""
     return all(x not in data for x in glob.glob(src))
 
 
 def check_command_installed(cmd):
+    """Try to call selected cmd and return True. If command not installed to system raise unittest.SkipTest"""
     try:
         subprocess.call(cmd, stderr=subprocess.DEVNULL, stdout=subprocess.DEVNULL)
         return True
@@ -51,40 +53,12 @@ def check_command_installed(cmd):
 
 
 def check_command_usage(cmd):
+    """Check that selected command output contain word 'usage'."""
     # check call without arguments
     c = subprocess.Popen(cmd, shell=True, stderr=subprocess.PIPE)
     return b'usage' in c.communicate()[1]
 
 
 def stdout_to_list(stdout):
+    """Convert stdout (str) to list of stripped strings"""
     return [x.strip() for x in stdout.split('\n') if x.strip()]
-
-
-class SystemOperationTest:
-    __meta__ = ABCMeta
-
-    @property
-    @abstractmethod
-    def cmd(self): pass
-
-    @property
-    @abstractmethod
-    def src_data_dir(self): pass
-
-    @property
-    @abstractmethod
-    def dst_data_dir(self): pass
-
-    def test_glob_command(self):
-        # Package must be installed to system!
-        # check command available in system
-        check_command_installed(self.cmd)
-        for glb in GLOBS:
-            generate_data(self.src_data_dir)
-            data = glob.glob(join(self.src_data_dir, glb))
-            subprocess.call('{} "{}" {}'.format(self.cmd, join(self.src_data_dir, glb), self.dst_data_dir), shell=True)
-            self.assertTrue(check_glob(self.src_data_dir, self.dst_data_dir, data))
-            clean(self.dst_data_dir)
-
-        # check call without arguments
-        self.assertTrue(check_command_usage(self.cmd))
